Traceur: Remove the disabled icon and clean it up.
The disabled icon was actually the same as the enabled icon! Also, we
didn't need both of them, since the tile state will change the icon
color for us.

Bug: 37281960
Test: Compiled, ran, saw the tile icon look correct while tracing
(white) and while not tracing (grayed out). Also saw the tile icon in
other places that it should be, like popup dialogs.

Change-Id: I4526b1aec25e0c274bc89dd23071dc7d60813143