Traceur: UI simplification and cleanup.
Removed the popups that appear when you're interacting with Traceur via
the tile, so that when you tap the tile tracing immediately begins, and
when you tap it again the trace is automatically stopped, dumped to
file, and the sharing notification appears. The text on the tile will
toggle between "Start tracing" and "Tracing".

Bug: 37281960
Test: The UX works as described above. Sometimes the active/inactive
state of the tile doesn't update immediately.

Change-Id: I50c130b6e682dd2769b5fdf74627d8a48895f1c2