Traceur: Remove the manual scripts for setting tracing permissions.
We'll be allowing all tracepoints for userdebug builds specifically for
Traceur users, so we shouldn't need these scripts and the popups
explaining how to run them any more.

I've left in the popup for if someone is running a trace and one of the
permissions isn't found, just in case something weird happens in the
future.

Bug: 37281960
Test: Built and installed Traceur, collected a trace, didn't see any
popups or app issues.

Change-Id: Iabbee41fabf1022f5073cb0f31547c99728750c5