Traceur: remove binder_driver from list of default tracepoints.
Due to b/37250231, binder_driver is corrupting traces. Remove this from
the defaults so the traces will be more likely to be successful.

In order to reset the defaults for all users of the app, change the
preferences key for tracepoints.

Refactor the preferences keys so that we only have to change the string
in one place in the future.

Bug: 37994014
Test: Opened Traceur app before this update and saw that binder_driver
was selected in the default set of tracepoints. Installed this change
(by using flashall) and saw that it was no longer in that list.
Test: Changed tracepoints, killed the app, and restarted it - the
tracepoint changes remained, indicating that they were being saved in
preferences. Similarly, started a trace, killed the app, and restarted
the app and saw that the trace was still noted as in progress.

Change-Id: I33e472b85529aa93266101f87bf3799246a4c326