Traceur: add camera and binder_driver to the list of default trace tags.
This change updates the list of default tags to add camera and
binder_driver, and changes the preference key for the current list so
that users' lists will update to the new defaults. If a user has changed
their tags, their changes will be lost.

This also changes the internal representation of the list of tags from
an ArrayMap to a TreeMap so they can easily be alphabetized and then
alphabetizes them.

Last, this change increases the app version, to make sure that OTA updates will
update the app to this version.

Bug: 62546628
Bug: 62382241
Test: Tested manually, including flashing an updated version of the app.
Change-Id: I3749da13bc09c6b21a563482edbd64093774ec4b