Traceur: Stop using the shell uid.
Traceur was previously running as shell, which was unintentionally
abusing selinux architecture. Now it gets its selinux permissions via
the traceur_app domain and no longer needs to share them with shell.

This change also updates the output directory for the traces to
/data/local/traces instead of /bugreports/, since without shell
permissions the app no longer has access to /bugreports/.

Last, this change adds a FileProvider to the manifest to allow us to
share the files from this new directory via email.

Bug: 68126425
Test: Enabled all tags, took a trace, and used
'adb pull /data/local/traces' to retrieve the trace files.
Test: Shared the trace files via gmail successfully.
Change-Id: Ibcb789354424276e94e3b2ea9451df26ab5e38eb