Traceur: Remove launcher icon, default Quick Settings tile to off.
Now that the Traceur app is available in User builds, it should only be
visible to the users if they've entered Developer Options. So this
change removes the launcher icon. Also, unless the user has explicitly
enabled the Quick Settings tile, it will default off.

The Traceur app is now only available via Developer Options, and if a
user has not explicitly enabled the Quick Settings tile, it will default
off.

Bug: 64762598
Test: Tests pass. Launcher icon no longer appears, but Traceur is
accessible via Settings.
Test: Quick Settings tile setting showed correct default behavior
and saved existing setting when there was one.

Change-Id: I500dfd71e9f1a6bf0dc05160c77974a0495e2fdd