Traceur: Use a different list of default tracepoints for user builds.
The default set of tracepoints included a few that are unavailable on
user builds due to permissions settings. We show an error when the user
requests tracepoints that aren't available (so they know they're not
getting what they expected), so we should populate a list of defaults
for user builds that won't trigger this error.

Bug: 72504322
Test: Tests pass. Userdebug builds populate the same default tracepoints
as before. User builds populate the new list of default tracepoints, and
there's no dialog when tracing indicating that some tracepoints are
unavailable.

Change-Id: I185df8de258473ace792cf3f8a9cb2a01217418f