Traceur: Remove Save button, ensure trace state is always correct.
This change removes the Save button and saves the trace every time.

It also makes sure that the Quickswitch tile and the Traceur activity UI
are always updated to represent the correct, current state of tracing.

Bug: 73509238
Test: Tests pass. Toggling the Trace button from the UI causes the QS
button to toggle. Toggling the QS button causes the Trace button in the
UI to toggle. When opening the app, it opens with the toggle in the
correct location.
Change-Id: I81e16fc45578f7074d0368d3ba0c5a410c655ad1