Traceur: Make sure QS Tile state is correct on first boot.
This changes the QS service to default to disabled, and also checks and
resets the service state on boot in addition to when the preference is
manually set. This should prevent edge cases where the QS tile is
available on first boot, when it shouldn't be.

This also adds a check to not show the QS tile when Developer Options
isn't visible.

Bug: 73087666
Bug: 73523779
Test: Tests pass.
Test: Enabled the QS tile setting, saw the QS tile visible. Switched to
a secondary user and did not see the QS tile. Switched back to primary
user and saw the QS tile.
Test: Cleared data on the device, did not see the QS tile on reboot. QS
tile appeared when setting was enabled.
Test: With the QS tile setting enabled in the app:
Disabled Developer Options, did not see the QS tile.
Reenabled Developer Options and the tile was visible.
Change-Id: I4ee473654bbb2a81a0d43f5667b64b43bfa8af59