Traceur: Add/remove icon from QS favorites when toggling the setting.
The existing code adds the Traceur icon to Quick Settings, but you
still have to edit your Quick Settings panel to move the icon to the place
you want. This behavior is different from the behavior for other
Developer Options QS tiles, which put the icon directly into your
favorites. That's much more user-friendly, so this change brings that
behavior into Traceur as well.

I also inexplicably called Quick Settings "Quickswitch" in a lot of this
code and so this change fixes that.

Bug: 74056507
Test: See the Traceur tile in the panel immediately when I enable
"Show Quick Settings tile." See it disappear immediately when I disable
it.
Test: Tests pass.

Change-Id: Iadbd81ccddbb9eb160e6fb2ea0f28e9246b75497