Traceur: Add StorageProvider so /data/local/traces appears in Files app.
This adds a StorageProvider, which allows us to browse the trace files
in Files apps and delete them from the device.

Note that this shortcut to the tracefiles directory will only be visible
in the Files apps if Developer Options is enabled.

Bug: 77220613
Test: With Developer Options enabled, reboot the phone. Open Files app
and see the new Storage Provider: "System traces" folder in the sidebar.
Test: Disable Developer Options and open Files app. Provider is not present.
Test: With Developer Options disabled, reboot the phone. Open Files app
and verify that the storage provider is not present.
Test: Enable Developer Options and open Files app. Provider is present.
Test: Enabled Developer Options and Quick Settings tile. Disabled
Developer Options and saw the Quick Settings tile disappear. Enabled
Developer Options and the Quick Settings tile did not reappear

Change-Id: I6c62250b04611eadcc4ddbd26023ad417ba0fdac