Traceur: Simplify proguard settings.
I shouldn't have included the LOCAL_PROGUARD_ENABLED := custom line in
Android.mk. That was keeping the generated proguard_options files from
being included automatically, requiring extra proguard rules to keep up.

This change removes the custom setting as well as the extra proguard
rules, keeping only support for the @Keep annotation and constructors
for classes that may be inflated from XML. The latter is required for
the app to start without crashing.

Bug: 76228127
Test: Manual test; app starts.

Change-Id: I207693afcfb2bc6b5a3a7ea4cb9844b6e604b995