Traceur: Move trace commands into a service.
This change is intended to reduce the likelihood that starting,
stopping, or saving a trace is interrupted due to an associated process
being killed. In order to prevent that, we've moved those activities
into a service.

The service runs in the foreground when performing atrace operations,
and will restart if it is stopped by the system.

Bug: 63867954
Test: atest TraceurUiTests - all tests passed.
Test: Added debugging code to sleep before dumping the trace and killed
the Traceur process while the trace was sleeping/dumping. Saw that the
AtraceService process restarted and the trace was dumped successfully
the second time.
Change-Id: I64391c4abe580d0926606f3a3a260bd1f123f6e2