Traceur: Handle situations where atrace is killed.
This change adds error handling so that if an atrace invocation is
killed or has an error when the trace is starting or dumping, Traceur
handles it slightly better.

If atrace is killed while the trace is starting, we'll stop the trace
and reset the tracking preference to "not tracing". This will appear to
the user like pressing the button had no effect.

If atrace is killed while dumping the trace, we'll abort dumping the
trace and not show the dialog for sharing. The partially dumped trace
is not cleaned up (but if the user selects "Clear saved traces" it will
be. This trace may or may not have usable content.

Note that if traceur itself is killed while the trace is starting or
being dumped, the AtraceService will be restarted, and the tracing
or dumping will restart with no interruption to the user. Therefore,
there is no specific error handling code for those cases.

Bug: 63867954
Test: Manually killed atrace during both cases and verified the
behavior.

Change-Id: I5ea59398b5169b793c589e38e81f727421a04cbb