Traceur: Turn QS tile off permanently when disabling Developer Options.
Previously, the preference was not unset when Developer Options was
disabled, which meant that the tile would return when Developer Options
was reenabled. This change unsets the preference so that the tile will
be off until the user explicitly requests it to come back.

Additionally, the ContentObserver is removed when the QS tile is not set
and added again when it is set, to lower the overhead of Traceur
generally.

Bug: 78652607
Test: atest TraceurUiTests
Test: Turned on Developer Options, toggled the preference, and saw the
tile appear. Turned off Developer Options and saw the tile disappear.
Turned it back on again and the tile did not reappear. Toggled the
preference again and the tile then did reappear.

Change-Id: I5443858093ec14b984afa6b40307ac3a47400260