Traceur: Add notification channels for ideal notification behavior.
Traceur notifications should peek onto the screen so that Traceur
users are clearly prompted for the next steps of taking, saving, and
sharing a trace. Therefore, Traceur needs notification channels to
ensure that this will happen all the time.

This also updates the min and target sdk versions in the manifest, which
provides a better default notification experience if people open
notification settings for Traceur.

Bug: 79114699
Test: Take traces using the app and the Quick Settings tile. See that
the notifications all peek onto the screen and vibrate with the default
vibration pattern in both cases.
Test: atest TraceurUiTests

Change-Id: Ib52f92e65c4233edbfa22be3627657187479d9fa