Traceur: Improve tests to reduce flakes.
This change makes two improvements to the Traceur tests:
1 - Adds automation to bypass the lockscreen if the test happens to
start on the lockscreen. This doesn't happen every time, but seems to
happen occasionally.

2 - Specifically waits for the popover notification to automatically
clear, then taps the notification from the shade. This prevents a race
condition where the popover notification is identified but disappears
before it can be tapped.

Bug: 77474475
Test: Ran the tests locally several times and experienced no local
flakes.
Test: Ran the tests starting from the lock screen and screen off.

Change-Id: Icecfa97091238a9245feaa76c4cd20f70fc888c7