Add a privacy warning popup in line with the trace share flow.
Since trace data could be considered sensitive in the same way that
bugreport data is, we should have similar warnings to users when they
share traces.

This change adds a dialog popup that appears after a user taps the
"Trace saved / Tap to share your trace" notification. The dialog has
similar wording to the Bug Report dialog and includes a "Don't show
again" checkbox which defaults to unchecked. When the user taps 'OK' on
this dialog, the share sheet appears (or the default share action
occurs).

If the user taps Cancel or backs out of the notification, the trace is
not shared, but it is still saved on the device in case the user wishes
to share it in the future. The user can delete the trace using the
existing UI in the System Tracing main app.

Bug: 116273848
Test: atest TraceurUiTests pass
Test: Manually tested several trace flows, including checking the box
and confirming that the dialog is not shown again but the share flow
continues successfully after the box is checked.

Change-Id: Ia91407284dd40adf0e9ccac850990c1b4543f71b