Add long trace support to Traceur
This change adds support for long traces in Traceur. including adding
additional UI parameters for the size and duration of long traces. These
parameters will be passed to Perfetto, and the trace will continue until
it meets either the size or the duration limit.

Perfetto will communicate back to Traceur that the trace is completed
using the setprop command defined in the traceur.rc file. The selinux
changes required to process that command are in the other commit in this
topic.

Note that we are exporting a separate service specifically for stopping
the trace, so that starting a trace is not also exported outside of the
Traceur app.

Bug: 116754134
Test: manual. Started a trace and stopped it via the quick settings
tile as well as via Perfetto ending the trace manually. In both cases
verified that the UX of the notifications worked correctly.

Change-Id: I309e699fb243396a56ac2f6edf4b617266ef8e4a