Ensure that the traced service is enabled before running Perfetto.
All devices are required to have Perfetto (via CTS testing),
but they are not required to enable it by default. This change
makes sure that the traced service is enabled when the user
takes a Perfetto trace using Traceur.

The service stays enabled subsequently, which is a ~1.5MB
RSS impact, because of the risk of potentially accidentally disabling it
on a device where it's supposed to be on.

Bug: 130784724
Test: manual
Change-Id: I7f85c57e6aa50a0c19806ac95814bde3b9a5e36a