Update validation to allow numbers in trace categories.
There can be numbers in trace categories, but we currently don't allow
them due to the data validation in Perfetto traces. This meant that i2c
tracepoints didn't work.

Also add capital letters to validation.

Bug: 133186121
Test: Started a perfetto trace with all current trace categories enabled
(including i2c) and did not see the "attempting to use an invalid tag"
error.

Change-Id: Ic4e23a2d6e36b7bba18878ceee2e94522f302d6e