Unregister the SharedPreferenceListener when the MainFragment pauses.
This is now getting triggered when the MainFragment isn't visible on the
screen, which results in a NPE when we try to getContext from this
listener.

Also, switch to using onStop/Start instead of onPause/Resume because
that should be a little more appropriate.

Bug: 133259473
Test: atest TraceurUiTests passes 10 times in a row. Manual testing.

Merged-In: I18fe4b5c788b4a43bf1a16bf6b6d3cbae78df590
Change-Id: I18fe4b5c788b4a43bf1a16bf6b6d3cbae78df590