Add a second notification channel.
This allows users to minimize the "Trace is being recorded"
notification, so that it's more subtle and less annoying to have around
for long periods of time. Because it's not ideal to let people swipe
away that notification completely lest they forget they're tracing, this
is hopefully a good compromise.

The notification channel named "Trace is being recorded" applies to that
notification, and the one named "Saving trace" applies to all of the
other notifications. That's because "Saving trace" is the most important
of the other group, and I don't currently have a better string
available.

Bug: 132918079
Test: Manual, changing the notifications and ensuring that they still
appeared.

Change-Id: Ibb1ad5fa4e54fef35a607efae75012ffdbf99eb2