Fix Perfetto trace buffer size to match atrace's.
The Traceur UI allows the user to pick a per-CPU buffer size. However,
we were passing this to Perfetto as the entire userspace buffer size.
This meant that when people switched from atrace to Perfetto traces,
their traces became much smaller.

Bug: 134191286
Test: Manually verified that the buffer size on a Wallleye is 8x what it
was without this change, by examining the logcat of the Perfetto config
output when the trace begins.

Change-Id: Ibb3cecb3f2b246c9ebcab65495ad1c0fbe4c26ed