Traceur: Automatically clear old traces.
Utilizes FileUtils.deleteOlderFiles to clear out older traces
when a new trace is taken.

Defaults to keeping at least three traces, and any traces modified
within the past month.

Bug: 115134821
Test: Take traces until four or more exist. Use touch
to set two of the traces' last modified dates to older
than a month. Take a trace and check that the older traces are
removed.
Test: Place two traces in /data/local/traces, and modify one or
more of their last modified dates to older than a month. Take
a trace. Check that the older traces are not removed.

Change-Id: If289e7a05be3f1d306479cee50743716c43109ea