Add developer options check for launching Traceur MainActivity
Previously, it was possible for apps to launch the System Tracing
app even when developer options was not enabled.

This change introduces a check at activity start that will close
the activity if developer options are not enabled.

Bug: 142936525
Test: Flashed phone with qt-qpr1-dev and verified that Traceur could
be launched from the Settings app and via the quick settings tile
long press.
Test: Verified that an app calling startActivity on the Traceur
MainActivity would fail if developer options were disabled.

Change-Id: I8b2da8cf732edf5c44b0a7f85f250752ae7d190f