Ensure trace files are always deletable by primary user.
When a System Tracing app tracing session is closed by a secondary
profile, the resultant trace file is then owned by the secondary
profile. This prevents users from deleting the file through the Files
app on the primary profile. This CL expands the file write permissions
to enable that use-case.

Test: Created a trace via work profile, and verified that it was not
deletable from the Files app w/out the patch, and was deletable with.
Bug: 135107222

Change-Id: I0b410a0c3a9e4a42211d26af784af818b5b3710f