Fix duplication of saving trace notifications in Traceur.
When in long trace mode, Traceur was simultaneously triggering
traceStop through its internal receiver and the external
StopTraceService. Due to a change in FGS notification handling,
both of these notifications hung around for ~5s.

This CL forces each cancellation route to use the same notification id.

Bug: 144948855
Test: Enabled long trace mode and took a trace. Verified traceStop
was called twice, and produced 2 notifications without change and 1 with.

Change-Id: I83d056533ba6edf4b33eb1445319f1e802b386c5