Read stdout and stderr in Traceur PerfettoUtils.traceStop.
In the bug referenced, the exec initiated process in traceStop appears
to hang, causing the saving trace notification to stick around
indefinitely. A common cause for this type of hang when using
RUNTIME.exec is the stdout or stderr pipe becoming full. This change
uses the AtraceUtils Logger class to constantly read from this pipes.

Bug: 137055264
Test: Built and took a trace, verified the trace saved correctly and
stderr was redirected to logcat.

Change-Id: I9187106a462fb663a566ea764a96d4763c47ae92
Merged-In: Iccd95ea73b6af2ec0f524f135d62189dc99a6d06