Initialize developer options ContentObserver at app start
Previously we initialized the ContentObserver in Receiver on receipt of
the BOOT_COMPLETE signal, but the process that handles that signal now
dies during boot. This change initializes the ContentObserver at
application start to ensure we always have an observer in place.

Additionally, we add the ContentObserver when the quick settings tile
service starts. This ensures that even if we do not open Traceur before
disabling developer options, the tile will be removed.

Bug: 150612980
Test: Traceur QS tile disappears after disabling developer options
Merged-In: I8512d9d1854490290c90b646f4cab2e939aa61c6
Change-Id: I8512d9d1854490290c90b646f4cab2e939aa61c6