Disable 'Attach traces to bugreport' on long trace
This change disables (grays out) the 'Attach traces to bug report'
option when a long trace is enabled. Long trace sessions cannot be
attached to bug reports because they are not entirely in memory, and
this change is intended to make this clear.

The user's setting for 'Attach traces...' is preserved.

Bug: 188898919
Test: Flashed with this change on both AOSP and internal to check that
        1) 'Attach traces...' is grayed out when long traces are enabled
	   on an internal build
	2) There is no visible change on AOSP, as 'Attach traces...'
	   does not exist
Change-Id: Ia5dcd903c1b7eefa751c3e76dfdc7fc3e5bab137
Merged-In: Ia5dcd903c1b7eefa751c3e76dfdc7fc3e5bab137